first version of refactored flair that passes all my original tests
