+++ bin/flair_diffSplice
@@ -1,0 +1,1 @@
+../src/flair/flair_diffSplice.py
