+++ test/test_expected/test.diffexp.genes_deseq2_MCF7_v_A549.tsv
@@ -1,16 +1,16 @@
 sample	baseMean	log2FoldChange	lfcSE	pvalue	padj
-ENSG00000087086.15	22427	-4.83	0.31	7.72e-55	1.02e-51
+ENSG00000087086.15	22427	-4.83	0.31	7.72000000000000e-55	1.02e-51
 ENSG00000160182.3	5785.11	10.91	0.84	4.2e-40	2.77e-37
 ENSG00000196154.12	1370.87	-10.96	0.97	7.57e-30	3.32e-27
 ENSG00000110492.16	341.36	6.77	0.63	6.23e-28	2.05e-25
 ENSG00000196542.8	450.37	14.03	3.26	2.86e-23	7.54e-21
-ENSG00000165092.13	903.39	-14.43	3.49	3.02e-22	6.64e-20
-ENSG00000135046.14	903.88	-14.41	3.49	2.16e-21	3.55e-19
+ENSG00000165092.13	903.39	-14.41	3.47	3.02e-22	6.64e-20
+ENSG00000135046.14	903.88	-14.4	3.48	2.16e-21	3.55e-19
 ENSG00000118640.11	279.99	4.24	0.46	1.99e-21	3.55e-19
 ENSG00000167815.12	473.65	10.04	1.13	1.08e-19	1.58e-17
 ENSG00000143320.9	166.96	6.09	0.72	1.07e-18	1.41e-16
 ENSG00000169908.12	379.3	-13.03	3.34	1.53e-18	1.83e-16
-ENSG00000137731.14	382.88	-13.04	3.34	1.74e-18	1.91e-16
+ENSG00000137731.14	382.88	-13.03	3.33	1.74e-18	1.91e-16
 ENSG00000111716.14	605.77	-11.14	1.43	2.14e-18	2.17e-16
 chr7:93807000	81	5.51	0.65	1.14e-17	1.08e-15
 ENSG00000187608.10	104.21	4.18	0.51	3.83e-17	3.37e-15
@@ -108,7 +108,7 @@ ENSG00000103490.14	72.92	2.19	0.6	4.15e-05	0.000516
 ENSG00000147592.9	123.5	-2.02	0.55	4.42e-05	0.000544
 ENSG00000188643.11	113.79	-2.82	0.8	4.49e-05	0.000549
 ENSG00000108518.8	2665.4	-1.29	0.34	5.56e-05	0.000672
-ENSG00000123144.11	214.01	1.37	0.37	6.41e-05	0.000768
+ENSG00000123144.11	214.01	1.38	0.37	6.41e-05	0.000768
 chr7:135674000	181.94	1.61	0.44	6.82e-05	0.00081
 chr14:37589000	123.6	3.09	0.94	8.34e-05	0.000981
 ENSG00000137309.20	66.21	-2.59	0.77	8.61e-05	0.000995
@@ -215,7 +215,7 @@ ENSG00000164167.12	88.1	1.2	0.48	0.00488	0.0302
 chr2:84905000	48.89	-1.35	0.55	0.00497	0.0306
 ENSG00000143153.13	64.35	-1.93	0.91	0.005	0.0307
 ENSG00000074696.13	56.36	-1.91	0.89	0.00503	0.0307
-ENSG00000120686.12	156.28	0.91	0.35	0.00528	0.032
+ENSG00000120686.12	156.28	0.92	0.35	0.00528	0.032
 ENSG00000120742.11	184.06	0.96	0.38	0.00582	0.0352
 chr16:21866000	64.64	2.61	1.65	0.00607	0.0364
 ENSG00000171159.5	131.55	1.49	0.65	0.00606	0.0364
