+++ test/test_expected/test.diffexp.isoforms_deseq2_MCF7_v_A549.tsv
@@ -1,13 +1,13 @@
 sample	baseMean	log2FoldChange	lfcSE	pvalue	padj
-ENST00000331825.11_ENSG00000087086.15	22282.77	-4.82	0.31	4.18e-57	6.17e-54
-ENST00000291527.3_ENSG00000160182.3	5768.95	10.92	0.84	1.1e-40	8.13e-38
-ENST00000368716.9_ENSG00000196154.12	1360.72	-10.95	0.96	1.81e-30	8.9e-28
+ENST00000331825.11_ENSG00000087086.15	22282.77	-4.82	0.31	4.18000000000000e-57	6.17000000000000e-54
+ENST00000291527.3_ENSG00000160182.3	5768.95	10.92	0.84	1.1e-40	8.13000000000000e-38
+ENST00000368716.9_ENSG00000196154.12	1360.72	-10.96	0.96	1.81e-30	8.9e-28
 ENST00000395566.9_ENSG00000110492.16	341.69	6.78	0.6	7.14e-30	2.63e-27
 ENST00000359175.8_ENSG00000196542.8	451.01	14.02	3.25	1.05e-23	3.11e-21
 ENST00000263864.10_ENSG00000118640.11	280	4.25	0.44	1.34e-23	3.3e-21
 ENST00000301522.3_ENSG00000167815.12	470.2	10.05	1.12	6.65e-20	1.09e-17
 ENST00000297785.8_ENSG00000165092.13	592.83	-13.74	3.42	6.43e-20	1.09e-17
-ENST00000257497.11_ENSG00000135046.14	785.15	-14.17	3.49	6.14e-20	1.09e-17
+ENST00000257497.11_ENSG00000135046.14	785.15	-14.17	3.48	6.14e-20	1.09e-17
 5834b38f-1003-4585-9284-7c0588a61574_chr7:93807000	81.36	5.51	0.63	7.25e-19	9.85e-17
 ENST00000368222.8_ENSG00000143320.9	168.94	6.11	0.72	7.34e-19	9.85e-17
 ENST00000305366.8_ENSG00000169908.12	377.3	-13.01	3.33	1.24e-18	1.4e-16
@@ -33,7 +33,7 @@ ENST00000375602.2_ENSG00000204379.11	83.38	-10.54	3.07	7.84e-13	3.73e-11
 ENST00000329305.6_ENSG00000198467.16	63.44	-10.1	2.99	1.11e-12	5.1e-11
 ENST00000370206.9-1_ENSG00000117519.16	72.72	-10.32	3.03	1.28e-12	5.71e-11
 5e3b558a-0f48-4e13-b9c6-0f4968ef0dde_chr6:151615000	71.26	10.95	3.07	2.25e-12	9.75e-11
-ENST00000296027.5-1_ENSG00000163735.7	223.87	-12.02	3.38	3.74e-12	1.57e-10
+ENST00000296027.5-1_ENSG00000163735.7	223.87	-12.03	3.38	3.74e-12	1.57e-10
 ENST00000451270.7_ENSG00000182718.18	510.09	-3.04	0.46	3.85e-12	1.58e-10
 1531fc1a-5817-42a6-9944-f53c01b609e9_ENSG00000165092.13	82.51	-10.49	3.09	4.05e-12	1.61e-10
 7b604afd-139e-470f-b1ec-170d8edd8f1b_chr8:119245000	162.95	12.19	3.31	6.27e-12	2.43e-10
@@ -71,7 +71,7 @@ ENST00000367889.8_ENSG00000143198.13	200.82	-2.62	0.46	1.92e-09	4.16e-08
 837e6f5d-7b93-4db8-a0d0-77ad3dacf56c_chr2:177538000	72.86	-10.14	3.19	2.92e-09	6.15e-08
 ENST00000341049.7_ENSG00000105974.13	53.54	-5.31	0.98	3.82e-09	7.93e-08
 ENST00000524595.5_ENSG00000150782.12	67.27	-7.03	1.24	4.02e-09	8.24e-08
-ENST00000389614.6_ENSG00000176153.12	1022.76	-3.57	0.66	4.7e-09	9.5e-08
+ENST00000389614.6_ENSG00000176153.12	1022.76	-3.56	0.66	4.7e-09	9.5e-08
 ENST00000458549.7_ENSG00000180694.14	271.03	4.13	0.76	4.93e-09	9.82e-08
 ENST00000299964.4_ENSG00000166741.8	80.79	-10.27	3.25	6.03e-09	1.19e-07
 67cd878d-0345-4788-b910-84a23aa68134_ENSG00000205426.10	114.16	-10.79	3.36	6.2e-09	1.2e-07
@@ -130,7 +130,7 @@ ENST00000305105.3_ENSG00000106400.12	271.37	1.39	0.35	2.49e-05	0.000286
 ENST00000677968.1_ENSG00000182718.18	88.43	-2.91	0.81	2.59e-05	0.000296
 ENST00000381342.7_ENSG00000125835.19	412.26	-1.57	0.4	2.61e-05	0.000297
 ENST00000301821.11_ENSG00000168028.14	1578.64	-1.4	0.35	2.72e-05	0.000306
-ENST00000343986.9_ENSG00000171858.18	5658.43	1.06	0.27	3.05e-05	0.000341
+ENST00000343986.9_ENSG00000171858.18	5658.43	1.07	0.27	3.05e-05	0.000341
 ENST00000368706.9_ENSG00000188643.11	113.49	-2.82	0.79	3.99e-05	0.000443
 ENST00000336458.13_ENSG00000132507.18	404.94	-1.51	0.4	4.4e-05	0.000484
 ENST00000361427.6_ENSG00000198830.11	874.53	1.51	0.4	4.46e-05	0.000488
@@ -243,7 +243,7 @@ ENST00000229595.6_ENSG00000111875.8	81.78	1.07	0.4	0.00338	0.0207
 b1c1cc53-6f17-4844-bd6b-46a3e93bc22c_chr3:10152000	58.53	1.87	0.81	0.00366	0.0223
 9004d520-c0ee-4d2f-a2af-422a9062b1d7_chr17:59947000	172.84	2.73	1.49	0.00383	0.0233
 ENST00000647248.2_ENSG00000182899.17	3921.69	1.09	0.41	0.00387	0.0234
-24814dad-d716-4e3c-861e-bd690b9253d9_chr5:72205000	101.78	-3.32	3.17	0.00391	0.0235
+24814dad-d716-4e3c-861e-bd690b9253d9_chr5:72205000	101.78	-3.33	3.16	0.00391	0.0235
 ENST00000261875.10_ENSG00000074696.13	56.54	-1.95	0.88	0.00403	0.0242
 ENST00000376510.8_ENSG00000148450.13	57.72	1.5	0.62	0.00406	0.0242
 4a77171d-8352-4d4f-aa64-feb0902b5fac_ENSG00000160345.13	51.51	1.66	0.7	0.00409	0.0243
